+++ flutter_ffi_plugin/example/native/hub/src/bridge/interface_web.rs
@@ -76,9 +76,13 @@ pub fn prepare_channels_extern() {
 #[wasm_bindgen]
 extern "C" {
     #[wasm_bindgen(js_name = rinf_send_rust_signal_extern)]
-    pub fn send_rust_signal_extern(resource: i32, message_raw: Uint8Array, blob_raw: Uint8Array);
+    pub fn send_rust_signal_extern_raw(
+        resource: i32,
+        message_raw: Uint8Array,
+        blob_raw: Uint8Array,
+    );
     #[wasm_bindgen(js_name = rinf_respond_to_dart_extern)]
-    pub fn respond_to_dart_extern(
+    pub fn respond_to_dart_extern_raw(
         id: i32,
         successful: bool,
         message_raw: Uint8Array,
@@ -87,4 +91,36 @@ extern "C" {
     #[cfg(debug_assertions)]
     #[wasm_bindgen(js_name = rinf_send_rust_report_extern)]
     pub fn send_rust_report_extern(rust_report: String);
+}
+
+pub fn send_rust_signal_extern(rust_signal: RustSignal) {
+    let message_raw = rust_signal.message.unwrap_or(vec![]);
+    let blob_raw = rust_signal.blob.unwrap_or(vec![]);
+
+    send_rust_signal_extern_raw(
+        rust_signal.resource,
+        rinf::externs::js_sys::Uint8Array::from(message_raw.as_slice()),
+        rinf::externs::js_sys::Uint8Array::from(blob_raw.as_slice()),
+    );
+}
+
+pub fn respond_to_dart_extern(response_unique: RustResponseUnique) {
+    let option = response_unique.response;
+
+    let (successful, message_raw, blob_raw) = if let Some(rust_response) = option {
+        (
+            true,
+            rust_response.message.unwrap_or(vec![]),
+            rust_response.blob.unwrap_or(vec![]),
+        )
+    } else {
+        (false, vec![], vec![])
+    };
+
+    respond_to_dart_extern_raw(
+        response_unique.id,
+        successful,
+        rinf::externs::js_sys::Uint8Array::from(message_raw.as_slice()),
+        rinf::externs::js_sys::Uint8Array::from(blob_raw.as_slice()),
+    );
 }
